moved isRefType to a util module. Moved RefObjectModels out to live with it's other similar types.
